fastapi_app/pyproject.toml
[project]
name = "fastapi-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["fastapi[all]>=0.119.0", "paho-mqtt>=2.1.0"]

[tool.ruff]
select = ["I", "F"]
